re-organizing redux store
